[modules/led] Move board specific code out of modules
Move it to a new dir afc-common
