+++ port/board/CMakeLists.txt
@@ -4,6 +4,10 @@ if(NOT BOARD)
   message(FATAL_ERROR "${BoldRed}Target board was not selected! You must pass the flag -DBOARD=<board_name>${ColourReset}")
 endif()
 
+if(${BOARD} MATCHES "^(afc)")
+  add_subdirectory(afc-common)
+endif()
+
 if(${BOARD} MATCHES "^(afc-bpm)$")
 
   if(NOT VERSION)
